requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0

# Machine Learning
tensorflow==2.13.0
scikit-learn==1.3.0
optuna==3.3.0

# Technical Analysis
ta==0.10.2

# Data Processing
pytz==2023.3
joblib==1.3.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1

# Logging & Monitoring
rich==13.4.2

# Development Tools
jupyter==1.0.0
black==23.7.0
flake8==6.1.0

# Documentation
sphinx==7.1.1
sphinx-rtd-theme==1.2.2 